mobileapi/requirements.txt
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
Django==5.0.9
django-cors-headers==4.4.0
django-mssql-backend==2.8.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
mssql==1.0.1
mssql-django==1.5
packaging==24.1
PyJWT==2.9.0
pymssql==2.3.1
pyodbc==5.1.0
pytz==2024.2
SQLAlchemy==2.0.35
sqlparse==0.5.1
typing_extensions==4.12.2
Werkzeug==3.0.4
whitenoise==6.7.0
